Moved symbols wrap to SolidGaugeComposition file
