+++ ts/Series/SolidGauge/SolidGaugeSeries.ts
@@ -12,6 +12,12 @@
 
 'use strict';
 
+/* *
+ *
+ *  Imports
+ *
+ * */
+
 import type SolidGaugePoint from './SolidGaugePoint';
 import type SolidGaugeSeriesOptions from './SolidGaugeSeriesOptions';
 import type SVGAttributes from '../../Core/Renderer/SVG/SVGAttributes';
@@ -20,13 +26,12 @@ import SolidGaugeAxis from './SolidGaugeAxis.js';
 import BaseSeries from '../../Core/Series/Series.js';
 const {
     seriesTypes: {
-        gauge: GaugeSeries
+        gauge: GaugeSeries,
+        pie: {
+            prototype: pieProto
+        }
     }
 } = BaseSeries;
-import H from '../../Core/Globals.js';
-const {
-    Renderer
-} = H;
 import LegendSymbolMixin from '../../Mixins/LegendSymbol.js';
 import U from '../../Core/Utilities.js';
 const {
@@ -35,94 +40,10 @@ const {
     isNumber,
     merge,
     pick,
-    pInt,
-    wrap
+    pInt
 } = U;
 
-/**
- * Internal types
- * @private
- */
-declare global {
-    namespace Highcharts {
-        interface SymbolOptionsObject {
-            rounded?: boolean;
-        }
-    }
-}
-
-/**
- * Additional options, depending on the actual symbol drawn.
- *
- * @interface Highcharts.SymbolOptionsObject
- *//**
- * Whether to draw rounded edges.
- * @name Highcharts.SymbolOptionsObject#rounded
- * @type {boolean|undefined}
- */
-
-/**
- * Symbol definition of an arc with round edges.
- *
- * @private
- * @function Highcharts.Renderer#symbols.arc
- *
- * @param {number} x
- *        The X coordinate for the top left position.
- *
- * @param {number} y
- *        The Y coordinate for the top left position.
- *
- * @param {number} w
- *        The pixel width.
- *
- * @param {number} h
- *        The pixel height.
- *
- * @param {Highcharts.SymbolOptionsObject} [options]
- *        Additional options, depending on the actual symbol drawn.
- *
- * @return {Highcharts.SVGPathArray}
- *         Path of the created arc.
- */
-wrap(
-    Renderer.prototype.symbols,
-    'arc',
-    function (
-        proceed: Function,
-        x: number,
-        y: number,
-        w: number,
-        h: number,
-        options: Highcharts.SymbolOptionsObject
-    ): SVGPath {
-        var arc = proceed,
-            path: SVGPath = arc(x, y, w, h, options);
-
-        if (options.rounded) {
-            var r = options.r || w,
-                smallR = (r - (options.innerR || 0)) / 2,
-                outerArcStart = path[0],
-                innerArcStart = path[2];
-
-            if (outerArcStart[0] === 'M' && innerArcStart[0] === 'L') {
-                const x1 = outerArcStart[1],
-                    y1 = outerArcStart[2],
-                    x2 = innerArcStart[1],
-                    y2 = innerArcStart[2],
-                    roundStart: SVGPath.Arc = ['A', smallR, smallR, 0, 1, 1, x1, y1],
-                    roundEnd: SVGPath.Arc = ['A', smallR, smallR, 0, 1, 1, x2, y2];
-
-                // Replace the line segment and the last close segment
-                path[2] = roundEnd;
-                path[4] = roundStart;
-            }
-        }
-
-        return path;
-    }
-);
-
+import './SolidGaugeComposition.js';
 
 /**
  * A solid gauge is a circular gauge where the value is indicated by a filled
@@ -291,7 +212,7 @@ class SolidGaugeSeries extends GaugeSeries {
         axis.initStops(axis.options);
 
         // Generate points and inherit data label position
-        H.seriesTypes.gauge.prototype.translate.call(this);
+        GaugeSeries.prototype.translate.call(this);
     }
 
     // Draw the points where each point is one needle.
@@ -444,7 +365,7 @@ class SolidGaugeSeries extends GaugeSeries {
     public animate(init?: boolean): void {
         if (!init) {
             this.startAngleRad = this.thresholdAngleRad;
-            H.seriesTypes.pie.prototype.animate.call(this, init);
+            pieProto.animate.call(this, init);
         }
     }
 }
